requirements.txt
anyio==3.6.2
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.87.0
fastapi-jwt-auth==0.5.0
h11==0.14.0
httptools==0.5.0
idna==3.4
passlib==1.7.4
pydantic==1.10.2
PyJWT==1.7.1
python-dotenv==0.21.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.21.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
